+++ src/components/AddRestaurantModal/add-restaurant-modal.css
@@ -60,8 +60,7 @@
 }
 
 .form-item input,
-.form-item textarea,
-.form-item select {
+.form-item textarea {
   padding: 8px;
   margin: 6px 0;
 
@@ -73,17 +72,6 @@
 
 .form-item textarea {
   resize: none;
-}
-
-.form-item select {
-  height: 44px;
-
-  padding: 8px;
-
-  border: 1px solid var(--grey-200);
-  border-radius: 8px;
-
-  color: var(--grey-300);
 }
 
 input[name='name'],
