+++ src/components/RestaurantFilter/RestaurantFilter.js
@@ -1,5 +1,5 @@
 import Restaurant from '../Common/Restaurant/Restaurant';
-import Select from '../Common/Select';
+import Select from '../Common/Select/Select';
 import { $ } from '../../utils/dom';
 import { RULES, SELECT_FILTER_DATA } from '../../constants/rules';
