+++ src/styles/index.css
@@ -1,7 +1,8 @@
 @import url('../components/Header/header.css');
 @import url('../components/AddRestaurantModal/add-restaurant-modal.css');
 @import url('../components/RestaurantFilter/restaurant-filter.css');
-@import url('../components/Restaurant/restaurant.css');
+@import url('../components/Common/Restaurant/restaurant.css');
+@import url('../components/Common/Select/select.css');
 
 * {
   padding: 0;
